refactor: key 형식 수정
